Extract evaluation-related methods from Position.cs to Evaluation.cs, making Position a partial class.
