+++ src/Lynx/Model/GameState.cs
@@ -38,11 +38,11 @@ public readonly struct GameState
 
         EnPassant = position.EnPassant;
         Castle = position.Castle;
-        IncrementalEvalAccumulator = position._incrementalEvalAccumulator;
-        IncrementalPhaseAccumulator = position._incrementalPhaseAccumulator;
+        IncrementalEvalAccumulator = position.IncrementalEvalAccumulator;
+        IncrementalPhaseAccumulator = position.IncrementalPhaseAccumulator;
 
         // We also save a copy of _isIncrementalEval, so that current move doesn't affect 'sibling' moves exploration
-        IsIncrementalEval = position._isIncrementalEval;
+        IsIncrementalEval = position.IsIncrementalEval;
     }
 }
